+++ src/main/java/com/sandymandy/pleasurecraft/client/rendering/renderers/AbstractGirlRenderer.java
@@ -302,7 +302,6 @@ public abstract class AbstractGirlRenderer<T extends GirlEntityScene, R extends 
         }
 
         VertexConsumer targetBuffer = buffer;
-
         int targetColor = renderColor;
 
         // Still allow UV offset